+++ examples/pitch_plots/plot_textured_background.py
@@ -15,16 +15,17 @@ from mplsoccer.pitch import Pitch
 from mplsoccer.utils import add_image
 
 # opening the background image
-# pic by webtreats: https://www.flickr.com/photos/webtreatsetc/
-# available at: https://www.flickr.com/photos/webtreatsetc/5756834840
-IMAGE_URL = 'https://live.staticflickr.com/5065/5756834840_e31c559b26_c_d.jpg'
+# pic by NASA/Expedition 40 crew member
+# available at: https://commons.wikimedia.org/wiki/File:ISS-40_Thunderheads_near_Borneo.jpg
+IMAGE_URL = ('https://upload.wikimedia.org/wikipedia/commons/'
+             '1/1d/ISS-40_Thunderheads_near_Borneo.jpg')
 image = Image.open(urlopen(IMAGE_URL))
 
 pitch = Pitch(pitch_color='None')
 fig, ax = pitch.draw(tight_layout=False)
 # adding the image and the image credits
 ax_image = add_image(image, fig, left=0, bottom=0, width=1, height=1)
-ax.text(70, 75, 'cloud pic by\nwebtreats (flickr)', color='white')
+ax.text(50, 75, 'cloud pic by\nNASA/Expedition 40 crew member', color='white')
 # set the pitch to be plotted after the image
 # note these numbers can be anything like 0.5, 0.2 as long
 # as the image zorder is behind the pitch zorder